Split verify.go in multiple files
