+++ verify/types.go
@@ -1,0 +1,4 @@
+package verify
+
+// This file contains type definitions used across the verify package.
+// Types are defined in verify.go to maintain backward compatibility.
